arm32: Move the FPA allocator out of arm_reg_alloc.c
The register allocator no longer contains any FPA specific code.
