+++ arch/arm32/ROMakefile
@@ -2,7 +2,7 @@
 
 COMPONENT = arm32
 
-OBJS = arm2_div arm_core arm_dump arm_encode arm_fpa_dist arm_gen arm_keywords arm_int_dist arm_link arm_peephole arm_reg_alloc arm_sub_section arm_vm arm_walker fpa fpa_gen arm_mem arm_heap assembler arm_expression
+OBJS = arm2_div arm_core arm_dump arm_encode arm_fpa_dist arm_gen arm_keywords arm_int_dist arm_link arm_peephole arm_reg_alloc arm_sub_section arm_vm arm_walker fpa fpa_alloc fpa_gen arm_mem arm_heap assembler arm_expression
 
 CFLAGS ?= -Wxla -Otime
