+++ arch/arm32/fpa_alloc.h
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2018 Mark Ryan
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef SUBTILIS_FPA_ALLOC_H__
+#define SUBTILIS_FPA_ALLOC_H__
+
+#include "arm_walker.h"
+
+void subtilis_fpa_alloc_init_walker(subtlis_arm_walker_t *walker,
+				    void *user_data);
+
+#endif
