+++ arch/arm32/fpa_gen.c
@@ -20,6 +20,7 @@
 #include "../../common/error_codes.h"
 #include "arm_core.h"
 #include "arm_gen.h"
+#include "fpa_alloc.h"
 #include "fpa_gen.h"
 
 void subtilis_fpa_gen_movr(subtilis_ir_section_t *s, size_t start,
@@ -935,4 +936,5 @@ void subtilis_arm_fpa_if_init(subtilis_arm_fp_if_t *fp_if)
 	fp_if->is_fixed_fn = subtilis_fpa_is_fixed;
 	fp_if->init_dist_walker_fn = subtilis_init_fpa_dist_walker;
 	fp_if->init_used_walker_fn = subtilis_init_fpa_used_walker;
+	fp_if->init_real_alloc_fn = subtilis_fpa_alloc_init_walker;
 }
